replaced calls to store with hooks
