+++ src/tests/utils/test-utils.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { render as rtlRender } from '@testing-library/react';
 import { Provider } from 'react-redux';
 
-import store from '../../common/store.ts';
+import { store } from '../../common/store.ts';
 
 function render(ui) {
   function Wrapper({ children }) {
